refactor: Chat Detail Data Module to Message Data Module
