--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/di/RepositoryModule.kt
@@ -1,16 +1,0 @@
-package com.apeun.gidaechi.chatdetail.di
-
-import com.apeun.gidaechi.chatdetail.ChatDetailRepository
-import com.apeun.gidaechi.chatdetail.repository.ChatDetailRepositoryImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface RepositoryModule {
-
-    @Binds
-    fun provideChatDetailRepository(chatDetailRepositoryImpl: ChatDetailRepositoryImpl): ChatDetailRepository
-}
