--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/mapper/ChatDetailEmojiMapper.kt
@@ -1,13 +1,0 @@
-package com.apeun.gidaechi.chatdetail.mapper
-
-import com.apeun.gidaechi.chatdetail.model.emoji.ChatDetailEmojiModel
-import com.apeun.gidaechi.network.chatdetail.response.emoji.ChatDetailEmojiResponse
-
-
-internal fun ChatDetailEmojiResponse.toModel() =
-    ChatDetailEmojiModel(
-        type = type.toChatType(),
-        eventList = eventList,
-        messageId = messageId,
-        emojiId = emojiId
-    )
