--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/mapper/ChatDetailSubMapper.kt
@@ -1,11 +1,0 @@
-package com.apeun.gidaechi.chatdetail.mapper
-
-import com.apeun.gidaechi.chatdetail.model.sub.ChatDetailSubModel
-import com.apeun.gidaechi.network.chatdetail.response.sub.ChatDetailSubResponse
-
-
-internal fun ChatDetailSubResponse.toModel() =
-    ChatDetailSubModel(
-        type = type.toChatType(),
-        eventList = eventList,
-    )
