--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/mapper/ChatTypeMessageMapper.kt
@@ -1,11 +1,0 @@
-package com.apeun.gidaechi.chatdetail.mapper
-
-import com.apeun.gidaechi.chatdetail.model.ChatMessageType
-import com.apeun.gidaechi.chatdetail.model.ChatType
-
-internal fun String.toChatMessageType(): ChatMessageType =
-    when(this) {
-        "ALIVE" -> ChatMessageType.ALIVE
-        "DELETE" -> ChatMessageType.DELETE
-        else -> ChatMessageType.DELETE
-    }
