--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/ChatDetailTypeModel.kt
@@ -1,5 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model
-
-abstract class ChatDetailTypeModel {
-    abstract val type: ChatType
-}
