--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/ChatMessageType.kt
@@ -1,6 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model
-
-enum class ChatMessageType {
-    ALIVE,
-    DELETE
-}
