--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/emoji/ChatDetailEmojiModel.kt
@@ -1,11 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.emoji
-
-import com.apeun.gidaechi.chatdetail.model.ChatDetailTypeModel
-import com.apeun.gidaechi.chatdetail.model.ChatType
-
-data class ChatDetailEmojiModel(
-    override val type: ChatType,
-    val eventList: List<Int>,
-    val messageId: String,
-    val emojiId : Int
-): ChatDetailTypeModel()
