--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/message/ChatDetailMessageDeleteModel.kt
@@ -1,10 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.message
-
-import com.apeun.gidaechi.chatdetail.model.ChatDetailTypeModel
-import com.apeun.gidaechi.chatdetail.model.ChatType
-
-data class ChatDetailMessageDeleteModel(
-    override val type: ChatType,
-    val eventList: List<Int>,
-    val messageId: String
-): ChatDetailTypeModel()
