--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/message/ChatDetailMessageEmojiModel.kt
@@ -1,6 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.message
-
-data class ChatDetailMessageEmojiModel(
-    val emojiId: Int,
-    val userId: List<Int>
-)
