--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/message/ChatDetailMessageLoadModel.kt
@@ -1,6 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.message
-
-data class ChatDetailMessageLoadModel(
-    val firstMessageId: String?,
-    val messages: List<ChatDetailMessageModel>
-)
