--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/message/ChatDetailMessageUserModel.kt
@@ -1,7 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.message
-
-data class ChatDetailMessageUserModel(
-    val id: Int,
-    val name: String,
-    val profile: String?
-)
