--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/room/ChatRoomModel.kt
@@ -1,15 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.room
-
-import java.time.LocalDateTime
-
-data class ChatRoomModel(
-    val id: Int,
-    val type: ChatRoomType,
-    val chatName: String,
-    val containUserCnt: Int,
-    val chatRoomImg: String,
-    val createdAt: LocalDateTime,
-    val workspaceId: String,
-    val memberList: List<Int>,
-    val chatStatusEnum: ChatRoomStatusType
-)
