--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/room/ChatRoomStatusType.kt
@@ -1,6 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.room
-
-enum class ChatRoomStatusType {
-    ALIVE,
-    DELETE
-}
