--- data/chat-detail/src/main/java/com/apeun/gidaechi/chatdetail/model/sub/ChatDetailSubModel.kt
@@ -1,9 +1,0 @@
-package com.apeun.gidaechi.chatdetail.model.sub
-
-import com.apeun.gidaechi.chatdetail.model.ChatDetailTypeModel
-import com.apeun.gidaechi.chatdetail.model.ChatType
-
-data class ChatDetailSubModel(
-    override val type: ChatType,
-    val eventList: List<Int>
-): ChatDetailTypeModel()
