+++ data/message/src/main/AndroidManifest.xml
