+++ data/message/src/main/java/com/apeun/gidaechi/message/di/RepositoryModule.kt
@@ -1,0 +1,16 @@
+package com.apeun.gidaechi.message.di
+
+import com.apeun.gidaechi.message.MessageRepository
+import com.apeun.gidaechi.message.repository.MessageRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal interface RepositoryModule {
+
+    @Binds
+    fun provideChatDetailRepository(chatDetailRepositoryImpl: MessageRepositoryImpl): MessageRepository
+}
