+++ data/message/src/main/java/com/apeun/gidaechi/message/mapper/MessageEmojiMapper.kt
@@ -1,0 +1,13 @@
+package com.apeun.gidaechi.message.mapper
+
+import com.apeun.gidaechi.message.model.emoji.MessageEmojiModel
+import com.apeun.gidaechi.network.chatdetail.response.emoji.ChatDetailEmojiResponse
+
+
+internal fun ChatDetailEmojiResponse.toModel() =
+    MessageEmojiModel(
+        type = type.toMessageType(),
+        eventList = eventList,
+        messageId = messageId,
+        emojiId = emojiId
+    )
