+++ data/message/src/main/java/com/apeun/gidaechi/message/mapper/MessageRoomMapper.kt
@@ -1,0 +1,17 @@
+package com.apeun.gidaechi.message.mapper
+
+import com.apeun.gidaechi.message.model.room.MessageRoomStatusType
+import com.apeun.gidaechi.message.model.room.MessageRoomType
+
+
+internal fun String.toMessageRoomType() =
+    when(this) {
+        "PERSONAL" -> MessageRoomType.PERSONAL
+        else -> MessageRoomType.GROUP
+    }
+
+internal fun String.toMessageRoomStatus() =
+    when(this) {
+        "ALIVE" -> MessageRoomStatusType.ALIVE
+        else -> MessageRoomStatusType.DELETE
+    }
