+++ data/message/src/main/java/com/apeun/gidaechi/message/mapper/MessageSubMapper.kt
@@ -1,0 +1,11 @@
+package com.apeun.gidaechi.message.mapper
+
+import com.apeun.gidaechi.message.model.sub.MessageSubModel
+import com.apeun.gidaechi.network.chatdetail.response.sub.ChatDetailSubResponse
+
+
+internal fun ChatDetailSubResponse.toModel() =
+    MessageSubModel(
+        type = type.toMessageType(),
+        eventList = eventList,
+    )
