+++ data/message/src/main/java/com/apeun/gidaechi/message/model/MessageLifeType.kt
@@ -1,0 +1,6 @@
+package com.apeun.gidaechi.message.model
+
+enum class MessageLifeType {
+    ALIVE,
+    DELETE
+}
