+++ data/message/src/main/java/com/apeun/gidaechi/message/model/MessageTypeModel.kt
@@ -1,0 +1,5 @@
+package com.apeun.gidaechi.message.model
+
+abstract class MessageTypeModel {
+    abstract val type: MessageType
+}
