+++ data/message/src/main/java/com/apeun/gidaechi/message/model/message/MessageEmojiModel.kt
@@ -1,0 +1,6 @@
+package com.apeun.gidaechi.message.model.message
+
+data class MessageEmojiModel(
+    val emojiId: Int,
+    val userId: List<Int>
+)
