+++ data/message/src/main/java/com/apeun/gidaechi/message/model/message/MessageLoadModel.kt
@@ -1,0 +1,6 @@
+package com.apeun.gidaechi.message.model.message
+
+data class MessageLoadModel(
+    val firstMessageId: String?,
+    val messages: List<MessageMessageModel>
+)
