+++ data/message/src/main/java/com/apeun/gidaechi/message/model/message/MessageMessageDeleteModel.kt
@@ -1,0 +1,10 @@
+package com.apeun.gidaechi.message.model.message
+
+import com.apeun.gidaechi.message.model.MessageTypeModel
+import com.apeun.gidaechi.message.model.MessageType
+
+data class MessageMessageDeleteModel(
+    override val type: MessageType,
+    val eventList: List<Int>,
+    val messageId: String
+): MessageTypeModel()
