+++ data/message/src/main/java/com/apeun/gidaechi/message/model/message/MessageMessageModel.kt
@@ -1,0 +1,21 @@
+package com.apeun.gidaechi.message.model.message
+
+import com.apeun.gidaechi.message.model.MessageTypeModel
+import com.apeun.gidaechi.message.model.MessageLifeType
+import com.apeun.gidaechi.message.model.MessageType
+import java.time.LocalDateTime
+
+data class MessageMessageModel(
+    val id: String,
+    val chatRoomId: Int,
+    override val type: MessageType,
+    val author: MessageUserModel,
+    val message: String,
+    val emojiList: List<MessageEmojiModel>,
+    val mention: List<Int>,
+    val mentionAll: Boolean,
+    val timestamp: LocalDateTime,
+    val read: List<Int>,
+    val joined: List<Int>,
+    val messageStatus: MessageLifeType
+): MessageTypeModel()
