+++ data/message/src/main/java/com/apeun/gidaechi/message/model/message/MessageUserModel.kt
@@ -1,0 +1,7 @@
+package com.apeun.gidaechi.message.model.message
+
+data class MessageUserModel(
+    val id: Int,
+    val name: String,
+    val profile: String?
+)
