+++ data/message/src/main/java/com/apeun/gidaechi/message/model/room/MessageRoomModel.kt
@@ -1,0 +1,15 @@
+package com.apeun.gidaechi.message.model.room
+
+import java.time.LocalDateTime
+
+data class MessageRoomModel(
+    val id: Int,
+    val type: MessageRoomType,
+    val chatName: String,
+    val containUserCnt: Int,
+    val chatRoomImg: String,
+    val createdAt: LocalDateTime,
+    val workspaceId: String,
+    val memberList: List<Int>,
+    val chatStatusEnum: MessageRoomStatusType
+)
