+++ data/message/src/main/java/com/apeun/gidaechi/message/model/room/MessageRoomStatusType.kt
@@ -1,0 +1,6 @@
+package com.apeun.gidaechi.message.model.room
+
+enum class MessageRoomStatusType {
+    ALIVE,
+    DELETE
+}
