+++ feature-main/chat-datail/src/main/java/com/apeun/gidaechi/chatdatail/model/ChatDetailUiState.kt
@@ -1,17 +1,15 @@
 package com.apeun.gidaechi.chatdatail.model
 
-import com.apeun.gidaechi.chatdetail.model.ChatMessageType
-import com.apeun.gidaechi.chatdetail.model.ChatType
-import com.apeun.gidaechi.chatdetail.model.message.ChatDetailMessageEmojiModel
-import com.apeun.gidaechi.chatdetail.model.message.ChatDetailMessageModel
-import com.apeun.gidaechi.chatdetail.model.message.ChatDetailMessageUserModel
+import com.apeun.gidaechi.message.model.MessageLifeType
+import com.apeun.gidaechi.message.model.message.MessageEmojiModel
+import com.apeun.gidaechi.message.model.message.MessageUserModel
 import java.time.LocalDateTime
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
 data class ChatDetailUiState(
     val roomInfo: ChatRoomState? = null,
-    val userInfo: ChatDetailMessageUserModel? = null,
+    val userInfo: MessageUserModel? = null,
     val nowPage: Int = 0,
     val isInit: Boolean = false,
     val isLastPage: Boolean = false,
@@ -21,22 +19,22 @@ data class ChatDetailUiState(
 data class ChatRoomState(
     val id: Int,
     val roomName: String,
-    val members: ImmutableList<ChatDetailMessageUserModel> = persistentListOf(),
+    val members: ImmutableList<MessageUserModel> = persistentListOf(),
 )
 
 data class ChatDetailMessageState(
     val id: String,
     val chatRoomId: Int,
     val type: ChatDetailChatTypeState,
-    val author: ChatDetailMessageUserModel,
+    val author: MessageUserModel,
     val message: String,
-    val emojiList: List<ChatDetailMessageEmojiModel>,
+    val emojiList: List<MessageEmojiModel>,
     val mention: List<Int>,
     val mentionAll: Boolean,
     val timestamp: LocalDateTime,
     val read: List<Int>,
     val joined: List<Int>,
-    val messageStatus: ChatMessageType,
+    val messageStatus: MessageLifeType,
     val isFirst: Boolean,
     val isLast: Boolean,
     val isMe: Boolean
@@ -45,7 +43,7 @@ data class ChatDetailMessageState(
         id = "id",
         chatRoomId = chatRoomId,
         type = type,
-        author = ChatDetailMessageUserModel(0, "qwe", null),
+        author = MessageUserModel(0, "qwe", null),
         message = "",
         emojiList = emptyList(),
         mention = emptyList(),
@@ -53,7 +51,7 @@ data class ChatDetailMessageState(
         timestamp = timestamp,
         read = emptyList(),
         joined = emptyList(),
-        messageStatus = ChatMessageType.ALIVE,
+        messageStatus = MessageLifeType.ALIVE,
         isFirst = false,
         isLast = false,
         isMe = false
